+++ c/nondominated.h
@@ -157,21 +157,29 @@ find_nondominated_2d_helper_(const double * restrict points, size_t size,
     ASSUME(size >= 2);
     const double **p = generate_sorted_pp_2d(points, size);
     // returning size means "no dominated solution found".
-    size_t n_nondom = size, k = 0, j = 1;
+    size_t n_nondom = size, j = 1;
     // When compiling with -O3, GCC is able to create two versions of this loop
     // and move keep_weakly out.
+    const double * restrict pk = p[0];
     do {
-        if (p[k][0] > p[j][0]) {
-            k = j;
-        } else if (!keep_weakly || p[k][0] != p[j][0] || p[k][1] != p[j][1]) {
-            size_t pos_first_dom = (p[j] - points) / 2;
-            if (nondom == NULL) {
-                // In this context, it means "position of the first dominated solution found".
-                n_nondom = pos_first_dom;
-                goto early_end;
+        const double * restrict pj = p[j];
+        if (pk[0] > pj[0]) {
+            pk = pj;
+        } else {
+            const bool k_eq_j = (pk[0] == pj[0]) & (pk[1] == pj[1]);
+            if (!keep_weakly || likely(!k_eq_j)) {
+                if (k_eq_j && pj < pk) // Only the first duplicated point is kept.
+                    SWAP(pj, pk);
+
+                size_t pos_first_dom = (pj - points) / 2;
+                if (unlikely(nondom == NULL)) {
+                    // In this context, it means "position of the first dominated solution found".
+                    n_nondom = pos_first_dom;
+                    goto early_end;
+                }
+                nondom[pos_first_dom] = false;
+                n_nondom--;
             }
-            nondom[pos_first_dom] = false;
-            n_nondom--;
         }
         j++;
     } while (j < size);
@@ -187,7 +195,7 @@ early_end:
 
 */
 static inline size_t
-find_dominated_2d_(const double *points, size_t size, const bool keep_weakly)
+find_dominated_2d_(const double * restrict points, size_t size, const bool keep_weakly)
 {
     return find_nondominated_2d_helper_(points, size, /*nondom=*/NULL, keep_weakly);
 }
@@ -213,7 +221,12 @@ find_nondominated_set_2d_(const double * restrict points, size_t size,
 }
 
 
-/*
+static inline void
+print_point(const double * p) {
+    printf("%g %g %g", p[0], p[1], p[2]);
+}
+
+/**
    3D dimension-sweep algorithm by H. T. Kung, F. Luccio, and F. P. Preparata.
    On Finding the Maxima of a Set of Vectors. Journal of the ACM,
    22(4):469–476, 1975.
@@ -221,19 +234,12 @@ find_nondominated_set_2d_(const double * restrict points, size_t size,
    A different implementation is available from Duarte M. Dias, Alexandre
    D. Jesus, Luís Paquete, A software library for archiving nondominated
    points, GECCO 2021. https://github.com/TLDart/nondLib/blob/main/nondlib.hpp
-
-   Duplicated points may be removed in any order due to qsort not being stable.
 */
-
-static inline void
-print_point(const double * p) {
-    printf("%g %g %g", p[0], p[1], p[2]);
-}
 static inline size_t
 find_nondominated_set_3d_helper(const double * restrict points, size_t size,
                                 bool * restrict nondom, const bool keep_weakly)
 {
-    ASSUME (size >= 2);
+    ASSUME(size >= 2);
     const double **p = generate_sorted_pp_3d(points, size);
 
     avl_tree_t tree;
@@ -249,24 +255,24 @@ find_nondominated_set_3d_helper(const double * restrict points, size_t size,
 
     // In this context, size means "no dominated solution found".
     size_t pos_last_dom = size, n_nondom = size, j = 1;
-    const double * prev = p[0];
+    const double * restrict pk = p[0];
     do {
         const double * restrict pj = p[j];
         // printf("pj: "); print_point(pj); printf("\n");
         bool dominated;
-        if (prev[0] > pj[0] || prev[1] > pj[1]) {
+        if (pk[0] > pj[0] || pk[1] > pj[1]) {
             avl_node_t * nodeaux;
             int res = avl_search_closest(&tree, pj, &nodeaux);
             assert(res != 0);
             if (res > 0) { // nodeaux goes before pj
-                prev = nodeaux->item;
+                const double * restrict prev = nodeaux->item;
                 assert(prev[0] != sentinel[0]);
                 assert(prev[0] <= pj[0]);
                 dominated = prev[1] <= pj[1];
                 // printf("res > 0: prev: "); print_point(prev); printf("\n");
                 nodeaux = nodeaux->next;
             } else if (nodeaux->prev) {// nodeaux goes after pj, so move to the next one.
-                prev = nodeaux->prev->item;
+                const double * restrict prev = nodeaux->prev->item;
                 assert(prev[0] != sentinel[0]);
                 assert(prev[0] <= pj[0]);
                 dominated = prev[1] <= pj[1];
@@ -276,7 +282,7 @@ find_nondominated_set_3d_helper(const double * restrict points, size_t size,
             }
 
             if (!dominated) { // pj is NOT dominated by a point in the tree.
-                const double * point = nodeaux->item;
+                const double * restrict point = nodeaux->item;
                 assert(pj[0] <= point[0]);
                 // Delete everything in the tree that is dominated by pj.
                 while (pj[1] <= point[1]) {
@@ -296,30 +302,39 @@ find_nondominated_set_3d_helper(const double * restrict points, size_t size,
         } else {
             // Handle duplicates and points that are dominated by the immediate
             // previous one.
-            dominated = !keep_weakly // We don't keep duplicates;
-                // or it is not a duplicate, so it is non-weakly dominated;
-                || prev[0] != pj[0] || prev[1] != pj[1] || prev[2] != pj[2]
-                // or prev was dominated, then this one is also dominated.
-                || pos_last_dom == (size_t) ((prev - points) / 3);
+            const bool k_eq_j = (pk[0] == pj[0]) & (pk[1] == pj[1]) & (pk[2] == pj[2]);
+            dominated = !keep_weakly;
+            if (dominated) { // We don't keep duplicates;
+                if (unlikely(k_eq_j) && pj < pk) // Only the first duplicated point is kept.
+                    SWAP(pk, pj);
+            } else { // or it is not a duplicate, so it is non-weakly dominated;
+                dominated = likely(!k_eq_j)
+                    // or pk was dominated, then this one is also dominated.
+                    || pos_last_dom == (size_t) ((pk - points) / 3);
+            }
             // printf("dom by prev: "); print_point(prev); printf("\n");
         }
         if (dominated) { // pj is dominated by a point in the tree or by prev.
             /* Map the order in p[], which is sorted, to the original order in
                points. */
             pos_last_dom = (pj - points) / 3;
-            if (nondom == NULL)
+            if (unlikely(nondom == NULL)) {
+                // In this context, it means "position of the first dominated solution found".
+                n_nondom = pos_last_dom;
                 goto early_end;
+            }
             nondom[pos_last_dom] = false;
             n_nondom--;
+        } else {
+            pk = pj;
         }
-        prev = pj;
         j++;
     } while (j < size);
 
 early_end:
     free(tnodes);
     free(p);
-    return (nondom == NULL) ? pos_last_dom : n_nondom;
+    return n_nondom;
 }
 
 static inline size_t
@@ -341,52 +356,6 @@ find_nondominated_set_3d_(const double * restrict points, size_t size, bool * re
 }
 
 static inline size_t
-find_dominated_point_agree_(const double * restrict points, dimension_t dim, size_t size,
-                            const bool keep_weakly, const enum objs_agree_t agree,
-                            const signed char * restrict minmax)
-{
-    ASSUME(dim > 3);
-    ASSUME(agree == AGREE_MINIMISE || agree == AGREE_MAXIMISE || agree == AGREE_NONE);
-    ASSUME(minmax == NULL || agree != AGREE_NONE);
-    for (size_t k = 1; k < size; k++) {
-        for (size_t j = 0; j < k; j++) {
-            const double * restrict pk = points + k * dim;
-            const double * restrict pj = points + j * dim;
-            // Use unsigned instead of bool to allow auto-vectorization.
-            unsigned k_leq_j = true, j_leq_k = true;
-            if (agree == AGREE_NONE) {
-                for (dimension_t d = 0; d < dim; d++) {
-                    double cmp = minmax[d] * (pk[d] - pj[d]);
-                    k_leq_j &= (cmp >= 0.0);
-                    j_leq_k &= (cmp <= 0.0);
-                }
-            } else {
-                for (dimension_t d = 0; d < dim; d++) {
-                    double cmp = (agree == AGREE_MINIMISE) ? pj[d] - pk[d] : pk[d] - pj[d];
-                    k_leq_j &= (cmp >= 0.0);
-                    j_leq_k &= (cmp <= 0.0);
-                }
-            }
-            /* Only the last duplicated point is kept. The 2D and 3D versions
-               depend on the order generated by qsort, which is not stable.  */
-            // k is removed if it is weakly dominated by j (unless keep_weakly == FALSE).
-            unsigned dom_k = j_leq_k & !(keep_weakly & k_leq_j);
-            // j is removed if it is dominated by k.
-            unsigned dom_j = k_leq_j & (!j_leq_k);
-
-            /* As soon as j_leq_k and k_leq_j become false, neither k or j will
-               be removed, so skip the rest.  */
-            if (!(dom_k | dom_j)) continue;
-
-            assert(dom_k ^ dom_j); // At least one but not both can be removed.
-
-            return dom_k ? k : j;
-        }
-    }
-    return size;
-}
-
-static inline size_t
 find_nondominated_set_agree_(const double * restrict points, dimension_t dim, size_t size,
                              bool * restrict nondom, const bool keep_weakly,
                              const enum objs_agree_t agree, const signed char * restrict minmax)
@@ -395,13 +364,25 @@ find_nondominated_set_agree_(const double * restrict points, dimension_t dim, si
     ASSUME(agree == AGREE_MINIMISE || agree == AGREE_MAXIMISE || agree == AGREE_NONE);
     ASSUME(minmax == NULL || agree != AGREE_NONE);
     size_t new_size = size;
-    for (size_t k = 1; k < size; k++) {
-        for (size_t j = 0; j < k; j++) {
-            assert(nondom[k]);
-            if (!nondom[j]) continue;
+    size_t min_k = 0;
+    for (size_t j = 1; j < size; j++) {
+        const double * restrict pj = points + j * dim;
+        size_t k = min_k;
+        if (likely(nondom != NULL)) {
+            assert(nondom[j]);
+            while (!nondom[k])
+                k++;
+            min_k = k;
+            ASSUME(k < j);
+            ASSUME(nondom[k]);
+        }
+        for (; k < j; k++) {
+            if (likely(nondom != NULL)) {
+                assert(nondom[j]);
+                if (!nondom[k]) continue;
+            }
 
             const double * restrict pk = points + k * dim;
-            const double * restrict pj = points + j * dim;
             // Use unsigned instead of bool to allow auto-vectorization.
             unsigned k_leq_j = true, j_leq_k = true;
             if (agree == AGREE_NONE) {
@@ -417,12 +398,11 @@ find_nondominated_set_agree_(const double * restrict points, dimension_t dim, si
                     j_leq_k &= (cmp <= 0.0);
                 }
             }
-            /* Only the last duplicated point is kept. The 2D and 3D versions
-               depend on the order generated by qsort, which is not stable.  */
-            // k is removed if it is weakly dominated by j (unless keep_weakly == FALSE).
-            unsigned dom_k = j_leq_k & !(keep_weakly & k_leq_j);
-            // j is removed if it is dominated by k.
-            unsigned dom_j = k_leq_j & (!j_leq_k);
+            // k is removed if it is dominated by j.
+            unsigned dom_k = (!k_leq_j) & j_leq_k;
+            // j is removed if it is weakly dominated by k (unless keep_weakly).
+            // Only the first duplicated point is kept.
+            unsigned dom_j = !keep_weakly ? k_leq_j : (k_leq_j & (!j_leq_k));
 
             /* As soon as j_leq_k and k_leq_j become false, neither k or j will
                be removed, so skip the rest.  */
@@ -430,17 +410,29 @@ find_nondominated_set_agree_(const double * restrict points, dimension_t dim, si
 
             assert(dom_k ^ dom_j); // At least one but not both can be removed.
 
-            if (dom_k) {
-                assert(nondom[k]);
-                nondom[k] = false;
+            if (unlikely(nondom == NULL)) {
+                return dom_k ? k : j;
+            }
+            if (dom_j) {
+                assert(nondom[j]);
+                nondom[j] = false;
                 new_size--;
                 break;
             }
-            nondom[j] = false;
+            nondom[k] = false;
             new_size--;
         }
     }
     return new_size;
+}
+
+static inline size_t
+find_dominated_point_agree_(const double * restrict points, dimension_t dim, size_t size,
+                            const bool keep_weakly, const enum objs_agree_t agree,
+                            const signed char * restrict minmax)
+{
+    return find_nondominated_set_agree_(points, dim, size, /* nondom=*/NULL,
+                                        keep_weakly, agree, minmax);
 }
 
 /* Stop as soon as one dominated point is found and return its position.
@@ -496,12 +488,13 @@ find_nondominated_set_(const double * restrict points, dimension_t dim, size_t s
                        const signed char * restrict minmax, enum objs_agree_t agree,
                        bool * nondom, const bool keep_weakly)
 {
-    ASSUME(minmax != NULL);
-    ASSUME(nondom != NULL);
     if (size < 2)
         return size;
 
     ASSUME(dim >= 2);
+    ASSUME(minmax != NULL);
+    ASSUME(nondom != NULL);
+
     if (dim <= 3) {
         const double *pp = force_agree_minimize (points, dim, size, minmax, agree);
         size_t res;
@@ -519,6 +512,9 @@ find_nondominated_set_(const double * restrict points, dimension_t dim, size_t s
        already to agree on minimization/maximization.  */
     if (agree == AGREE_NONE)
         agree = check_all_minimize_maximize(minmax, dim);
+
+    ASSUME(minmax != NULL);
+    ASSUME(nondom != NULL);
 
     switch (agree) {
       case AGREE_NONE:
@@ -601,8 +597,7 @@ find_weak_nondominated_set(const double * restrict points, dimension_t dim, size
                            const signed char * restrict minmax,
                            bool * restrict nondom)
 {
-    ASSUME(dim >= 2);
-    ASSUME(dim <= 32);
+    ASSUME(dim >= 2 && dim <= 32);
     return find_nondominated_set_(points, dim, size, minmax, AGREE_NONE, nondom,
                                   /* keep_weakly = */true);
 }
@@ -611,8 +606,7 @@ static inline size_t
 find_weak_nondominated_set_minimise(const double * restrict points, dimension_t dim,
                                     size_t size, bool * restrict nondom)
 {
-    ASSUME(dim >= 2);
-    ASSUME(dim <= 32);
+    ASSUME(dim >= 2 && dim <= 32);
     const signed char * minmax = minmax_minimise(dim);
     ASSUME(minmax != NULL);
     size_t new_size = find_weak_nondominated_set(points, dim, size, minmax,
@@ -627,13 +621,11 @@ get_nondominated_set (double **pareto_set_p,
                       const double *points, int dim, size_t size,
                       const signed char *minmax)
 {
-    ASSUME(dim >= 2);
-    ASSUME(dim <= 32);
-    ASSUME(size > 0);
+    ASSUME(dim >= 2 && dim <= 32);
 
-    bool *nondom = nondom_init(size);
+    bool * nondom = nondom_init(size);
     size_t new_size = find_nondominated_set(points, (dimension_t) dim, size, minmax, nondom);
-    double *pareto_set = malloc(sizeof (double) * new_size * dim);
+    double * pareto_set = malloc(new_size * dim * sizeof(*pareto_set));
 
     if (new_size < size) {
         size_t n = 0, k = 0;
@@ -656,11 +648,10 @@ static inline size_t
 filter_dominated_set(double * restrict points, int dim, size_t size,
                      const signed char * restrict minmax)
 {
-    ASSUME(dim >= 2);
-    ASSUME(dim <= 32);
+    ASSUME(dim >= 2 && dim <= 32);
     ASSUME(size > 0);
 
-    bool *nondom = nondom_init(size);
+    bool * nondom = nondom_init(size);
     size_t new_size = find_nondominated_set(points, (dimension_t) dim, size, minmax, nondom);
 
     if (new_size < size) {
@@ -684,8 +675,7 @@ is_nondominated_minmax(const double * restrict data, dimension_t nobj, size_t np
                        const signed char * restrict minmax, bool keep_weakly)
 {
     bool * nondom = nondom_init(npoint);
-    find_nondominated_set_(data, nobj, npoint, minmax, AGREE_NONE, nondom,
-                           /* keep_weakly = */keep_weakly);
+    find_nondominated_set_(data, nobj, npoint, minmax, AGREE_NONE, nondom, keep_weakly);
     return nondom;
 }